* libmutil: W32 port:   - Fixed warning messages   - Improved return value checking
git-svn-id: svn://svn.minisip.org/minisip/trunk@1819 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/source/IString.cxx
@@ -29,7 +29,7 @@ StringAtom::StringAtom(char *b, int len):buf(b),n(len){
 
 StringAtom::StringAtom(std::string s){
 	buf = strdup(s.c_str());
-	n = s.length();
+	n = (int)s.length();
 }
 
 StringAtom::~StringAtom(){
